[#109]영상찍을준비 완료
