+++ Fleming_App/New Folder/Localizable.xcstrings
@@ -196,6 +196,9 @@
     "실시간 소리 크기 그래프 (dB)" : {
 
     },
+    "한 번 더" : {
+
+    },
     "현재 소리 크기: %lld dB" : {
 
     },
